+++ Source/Krypton Components/Krypton.Toolkit/Palette Controls/KryptonColorTable365.cs
@@ -1,11 +1,11 @@
-ï»¿#region BSD License
+#region BSD License
 /*
  *
  * Original BSD 3-Clause License (https://github.com/ComponentFactory/Krypton/blob/master/LICENSE)
- *  Â© Component Factory Pty Ltd, 2006 - 2016, (Version 4.5.0.0) All rights reserved.
+ *  © Component Factory Pty Ltd, 2006 - 2016, (Version 4.5.0.0) All rights reserved.
  *
  *  New BSD 3-Clause License (https://github.com/Krypton-Suite/Standard-Toolkit/blob/master/LICENSE)
- *  Modifications by Peter Wagner (aka Wagnerp), Simon Coghlan (aka Smurf-IV), Giduac, Ahmed Abdelhameed, tobitege,  KamaniAR, Lesandro Gotardo (aka lesandrog), Jorge A. AvilÃ©s (aka mcpbcs) et al. 2017 - 2026. All rights reserved.
+ *  Modifications by Peter Wagner (aka Wagnerp), Simon Coghlan (aka Smurf-IV), Giduac, Ahmed Abdelhameed, tobitege,  KamaniAR, Lesandro Gotardo (aka lesandrog), Jorge A. Avilés (aka mcpbcs) et al. 2017 - 2026. All rights reserved.
  *
  */
 #endregion
@@ -18,15 +18,6 @@ namespace Krypton.Toolkit;
 public class KryptonColorTable365White : KryptonColorTable365
 {
     #region Identity
-    static KryptonColorTable365White()
-    {
-        // Get the font settings from the system
-        DefineFonts();
-
-        // We need to notice when system color settings change
-        SystemEvents.UserPreferenceChanged += OnUserPreferenceChanged;
-    }
-
     /// <summary>
     /// Initialize a new instance of the KryptonColorTable365White class.
     /// </summary>
@@ -65,8 +56,6 @@ public class KryptonColorTable365 : KryptonColorTable
     private static readonly Color _buttonCheckedEnd = Color.FromArgb(255, 216, 107);
     private static readonly Color _menuItemSelectedBegin = Color.FromArgb(251, 242, 215);
     private static readonly Color _menuItemSelectedEnd = Color.FromArgb(247, 224, 135);
-    private static Font _menuToolFont;
-    private static Font _statusFont;
     #endregion
 
     #region Instance Fields
@@ -75,22 +64,13 @@ public class KryptonColorTable365 : KryptonColorTable
     #endregion
 
     #region Identity
-    [SecuritySafeCritical]
-    static KryptonColorTable365()
-    {
-        // Get the font settings from the system
-        DefineFonts();
-
-        // We need to notice when system color settings change
-        SystemEvents.UserPreferenceChanged += OnUserPreferenceChanged;
-    }
-
     /// <summary>
     /// Initialize a new instance of the KryptonColorTable2010 class.
     /// </summary>
     /// <param name="colors">Source of </param>
     /// <param name="roundedEdges">Should have rounded edges.</param>
     /// <param name="palette">Associated palette instance.</param>
+    [SecuritySafeCritical]
     public KryptonColorTable365([DisallowNull] Color[] colors,
         InheritBool roundedEdges,
         PaletteBase palette)
@@ -552,30 +532,6 @@ public class KryptonColorTable365 : KryptonColorTable
     public override Color StatusStripText => _colors[(int)SchemeBaseColors.StatusStripText];
 
     #endregion
-
-    #region MenuStripFont
-    /// <summary>
-    /// Gets the font used on the menu strip.
-    /// </summary>
-    public override Font MenuStripFont => _menuToolFont;
-
-    #endregion
-
-    #region ToolStripFont
-    /// <summary>
-    /// Gets the font used on the tool strip.
-    /// </summary>
-    public override Font ToolStripFont => _menuToolFont;
-
-    #endregion
-
-    #region StatusStripFont
-    /// <summary>
-    /// Gets the font used on the status strip.
-    /// </summary>
-    public override Font StatusStripFont => _statusFont;
-
-    #endregion
     #endregion
 
     #region ToolStrip
@@ -650,20 +606,5 @@ public class KryptonColorTable365 : KryptonColorTable
     public override Color ToolStripPanelGradientEnd => _colors[(int)SchemeBaseColors.ToolStripBack];
 
     #endregion
-    #endregion
-
-    #region Implementation
-    protected static void DefineFonts()
-    {
-        // Create new font using system information
-        // TODO: Should be using base font
-        _menuToolFont = new Font(@"Segoe UI", SystemFonts.MenuFont!.SizeInPoints!, FontStyle.Regular);
-        _statusFont = new Font(@"Segoe UI", SystemFonts.StatusFont!.SizeInPoints!, FontStyle.Regular);
-    }
-
-    protected static void OnUserPreferenceChanged(object sender, UserPreferenceChangedEventArgs e) =>
-        // Update fonts to reflect any change in system settings
-        DefineFonts();
-
     #endregion
 }